# contracts/catalog-inventory-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Live Commerce catalog and inventory admin API",
    "version": "0.1.0",
    "description": "T04 merchant-only slice, not a complete SaaS/live-provider approval. Runtime requests derive Scope from a merchant session; client tenant headers and Host are not authority. No public reserve route."
  },
  "paths": {
    "/v1/admin/stores/{store_id}/catalog-ledger": {
      "get": {
        "operationId": "listCatalogLedger",
        "description": "Requires BOTH catalog:read and inventory:read in the same scoped transaction. q is a literal case-insensitive name/SKU substring (%, _, ! and backslash are literal); cursor binds store, warehouse, normalized q and status. Active warehouse required.",
        "security": [
          {
            "merchantBearer": []
          }
        ],
        "parameters": [
          {
            "name": "store_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "warehouse_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 120
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "active",
                "archived"
              ],
              "default": "all"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 1024
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scoped SKU-stock projection. No balance row returns zero values. Keyset cursor is not a snapshot.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "items",
                    "next_cursor"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "maxItems": 100,
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "required": [
                          "product_id",
                          "sku_id",
                          "warehouse_id",
                          "product_name",
                          "product_description",
                          "code",
                          "currency",
                          "status",
                          "price_minor",
                          "sku_version",
                          "on_hand",
                          "reserved",
                          "allocated",
                          "unavailable",
                          "available",
                          "balance_version",
                          "product_version",
                          "product_status",
                          "cover_image_id"
                        ],
                        "properties": {
                          "product_id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "sku_id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "warehouse_id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "product_name": {
                            "type": "string"
                          },
                          "product_description": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "currency": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "active",
                              "archived"
                            ]
                          },
                          "price_minor": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "sku_version": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "on_hand": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "reserved": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "allocated": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "unavailable": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "available": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "balance_version": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "product_version": {
                            "type": "integer",
                            "format": "int64",
                            "minimum": 1,
                            "description": "catalog-media: expected_version for PATCH/archive of the product"
                          },
                          "product_status": {
                            "enum": [
                              "active",
                              "archived"
                            ],
                            "description": "catalog-media: the product's own status (status also turns archived with its SKU)"
                          },
                          "cover_image_id": {
                            "type": [
                              "string",
                              "null"
                            ],
                            "format": "uuid",
                            "description": "catalog-media: id of the position-0 photo, null when none"
                          }
                        }
                      }
                    },
                    "next_cursor": {
                      "type": "string",
                      "maxLength": 1024
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "Server-generated request correlation ID"
              }
            }
          },
          "401": {
            "description": "Invalid merchant session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "Server-generated request correlation ID"
              }
            }
          },
          "403": {
            "description": "Visible store but operation not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "Server-generated request correlation ID"
              }
            }
          },
          "404": {
            "description": "Resource not visible in current scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "Server-generated request correlation ID"
              }
            }
          },
          "409": {
            "description": "Conflict, stale version or insufficient inventory",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "Server-generated request correlation ID"
              }
            }
          },
          "415": {
            "description": "application/json required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "Server-generated request correlation ID"
              }
            }
          },
          "422": {
            "description": "Semantic validation or query/cursor invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "Server-generated request correlation ID"
              }
            }
          },
          "500": {
            "description": "Internal failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "Server-generated request correlation ID"
              }
            }
          },
          "503": {
            "description": "Retry later with same key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                },
                "description": "Server-generated request correlation ID"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "merchantBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque short-lived merchant session; server stores hash only."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "request_id",
          "retryable",
          "details"
        ],
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 80
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string",
            "pattern": "^[0-9a-f]{32}$"
          },
          "retryable": {
            "type": "boolean"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
    }
  }
}